backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
validators==0.22.0
httpx==0.24.1
beautifulsoup4==4.12.2
supabase==1.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.19
psycopg2-binary==2.9.9
fastapi-cache2==0.2.1
redis==4.6.0
